requirements.txt
streamlit==1.31.1
scikit-learn==1.4.1
langdetect==1.0.9
transformers==4.37.2
torch>=2.0.0
python-docx==1.0.1
matplotlib==3.8.2
